break out ParquetFileWrapper into its own module
