+++ GCRCatalogs/dc2_truth_parquet.py
@@ -7,15 +7,14 @@ import os
 import re
 import warnings
 
-import pyarrow.parquet as pq
-import yaml
 from GCR import BaseGenericCatalog
+from .parquet import ParquetFileWrapper
 
 from .utils import first
 
 __all__ = ['DC2TruthParquetCatalog']
 
-class ParsePathInfo():
+class PathInfoParser():
     _group_pat = r'\{[a-zA-Z_][a-zA-Z0-9_]*\}'
     _known_ints = ('tract', 'visit', 'healpix')
     #_path_key = 'PATH'
@@ -78,85 +77,6 @@ class ParsePathInfo():
                 pass
         return d
 
-#from .dc2_dm_catalog import ParquetFileWrapper
-class ParquetFileWrapper():
-    def __init__(self, file_path, info=None):
-        self.path = file_path
-        self._handle = None
-        self._columns = None
-        self._info = info or dict()
-        self._schema = None
-        self._row_group = 0
-
-    @property
-    def handle(self):
-        if self._handle is None:
-            self._handle = pq.ParquetFile(self.path)
-        return self._handle
-
-    @property
-    def num_row_groups(self):
-        return self.handle.metadata.num_row_groups
-
-    @property
-    def row_group(self):
-        return self._row_group
-
-    @row_group.setter
-    def row_group(self, grp):
-        self._row_group = grp
-
-    def close(self):
-        self._handle = None
-
-    def __len__(self):
-        return int(self.handle.scan_contents)
-
-    def __contains__(self, item):
-        return item in self.columns
-
-    def read_columns(self, columns, as_dict=False):
-        d = self.handle.read(columns=columns).to_pandas()
-        if as_dict:
-            return {c: d[c].values for c in columns}
-        return d
-
-    def read_columns_row_group(self, columns, as_dict=False):
-        d = self.handle.read_row_group(self._row_group, columns=columns).to_pandas()
-        if as_dict:
-            return {c: d[c].values for c in columns}
-        return d
-        
-
-    @property
-    def info(self):
-        return dict(self._info)
-
-    def __getattr__(self, name):
-        if name not in self._info:
-            raise AttributeError('Attribute {} does not exist'.format(name))
-        return self._info[name]
-
-    @property
-    def columns(self):
-        if self._columns is None:
-            self._columns = [col for col in self.handle.schema.to_arrow_schema().names
-                             if re.match(r'__\w+__$', col) is None]
-        return list(self._columns)
-
-    @property
-    def native_schema(self):
-        if self._schema is None:
-            self._schema = {}
-            arrow_schema = self.handle.schema.to_arrow_schema()
-            for i in range(len(arrow_schema.names)):
-                tp = str(arrow_schema[i].type)
-                if tp == 'float': tp = 'float32'
-                else:
-                    if tp == 'double': tp = 'float42'
-                self._schema[arrow_schema.names[i]] = {'dtype' : tp}
-        
-        return self._schema
     
 class DC2TruthParquetCatalog(BaseGenericCatalog):
     r"""
@@ -168,25 +88,27 @@ class DC2TruthParquetCatalog(BaseGenericCatalog):
        Parameters
        ----------
        base_dir         (str): Directory of data files being served.  Required.
-       filename_pattern (str): Optional "enhanced regex" pattern of served data files.
+       filename_pattern (str): Optional "enhanced regex" pattern of served data
+                               files.
                                Default is match anything
 
-       If filename_pattern contains substrings like "{some_ident}"  where some_ident
-       is a legal identifier, this part of the pattern will be replaced with
-       a regex expression for a group matching a string of digits or word characters,
+       If filename_pattern contains substrings like "{some_ident}"  where 
+       some_ident is a legal identifier, this part of the pattern will be 
+       replaced with a regex expression for a group matching a string of 
+       digits or word characters,
        e.g.  
              (?P<some_ident>\d+) or              (?P<some_ident>\w+)
        The first form will be used iff the identifier is one of a well-known set
        with integer values, currently ('tract', 'visit', 'healpix')
 
        Such group names may be used subsequently as native_filter_quantities
-       If filename_pattern already includes standard regex syntax for named groups,
-       those group names may also be used as native filters
+       If filename_pattern already includes standard regex syntax for named 
+       groups, those group names may also be used as native filters
     """
 
     def _subclass_init(self, **kwargs):
         self.base_dir = kwargs['base_dir']
-        self.path_parser = ParsePathInfo(kwargs.get('filename_pattern','.*'))
+        self.path_parser = PathInfoParser(kwargs.get('filename_pattern','.*'))
 
         if not os.path.isdir(self.base_dir):
             raise ValueError('`base_dir` {} is not a valid directory'.format(self.base_dir))
@@ -206,15 +128,16 @@ class DC2TruthParquetCatalog(BaseGenericCatalog):
         """Return viable data sets from all files in self.base_dir
 
         Returns:
-            A list of  ParquetFileWrapper objects.  If any native filters come from
-            filepath re, dict of their values will be stored in the object
+            A list of  ParquetFileWrapper objects.  If any native filters come 
+            from filepath re, dict of their values will be stored in the object
         """
         datasets = list()
         for fname in sorted(os.listdir(self.base_dir)):
             info_dict = self.path_parser.file_info(fname)
             if info_dict == None:
                 continue
-            datasets.append(ParquetFileWrapper(os.path.join(self.base_dir, fname),
+            datasets.append(ParquetFileWrapper(os.path.join(self.base_dir,
+                                                            fname),
                                                info=info_dict))
         return datasets
 
@@ -232,12 +155,12 @@ class DC2TruthParquetCatalog(BaseGenericCatalog):
         return set(self._schema.keys()).union(self._native_filter_quantities)
 
     @staticmethod
-    def _obtain_native_data_dict(native_quantities_needed, native_quantity_getter):
+    def _obtain_native_data_dict(native_quantities_needed,
+                                 native_quantity_getter):
         """
         Overloading this so that we can query the database backend
         for multiple columns at once
         """
-        #return native_quantity_getter.read_columns(list(native_quantities_needed), as_dict=True)
         return native_quantity_getter.read_columns_row_group(list(native_quantities_needed), as_dict=True)
 
     def _iter_native_dataset(self, native_filters=None):
@@ -248,6 +171,3 @@ class DC2TruthParquetCatalog(BaseGenericCatalog):
             for i in range(dataset.num_row_groups):
                 dataset.row_group = i
                 yield dataset
-
-
-
